refactor - composable function in separate files
