+++ NFFrontend/app/src/main/java/com/example/nf_frontend/screens/homescreen/composant/TopBar.kt
@@ -1,0 +1,34 @@
+package com.example.nf_frontend.screens.homescreen.composant
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Person
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun TopBar() {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = "Mes cours",
+            style = MaterialTheme.typography.displayLarge,
+            modifier = Modifier.weight(1f)
+        )
+        IconButton(onClick = { /*TODO*/ }) {
+            Icon(Icons.Filled.Person, contentDescription = null)
+        }
+    }
+}
